question: abstract class instantiateable?
um, i would've previously called this instantiation, however S.O.
clearly says thats not the case.

+++ src/Nutrient.java
@@ -3,8 +3,8 @@
  */
 public abstract class Nutrient {
 
-    int grams;
-    Nutrient(int grams){
+    double grams;
+    Nutrient(double grams){
         this.grams = grams;
     }
